+++ CHANGES.rst
@@ -11,6 +11,9 @@ Changelog
 - add profile for Plone 6 and fix controlpanel icons.
   [petschki]
 
+- Use tox for testing, ensure tests are passing for Plone 5.0.x, 5.1.x and 5.2.x.
+  [thomasmassmann]
+
 
 1.0a1 (2016-10-06)
 ------------------